2 functions added, update & fix some bugs
The functions SMT, splitstring and stringToUint are made private, so they are internally called by the other functions. New functions: getTenderLenght and isPending to retrieve the number of total tenders published and to check whether a specific tender is still pending or not. Update the function for displaying the winner and getBidDetails.